refactor: useNotification, useProduct 분리
