+++ src/basic/hooks/index.ts
@@ -2,3 +2,5 @@ export * from './useCart';
 export * from './useCoupon';
 export * from './useLocalStorage';
 export * from './useCartItems';
+export * from './useNotification';
+export * from './useProduct';
